+++ packages/query/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ydbjs/query
 
+## 6.0.6
+
+### Patch Changes
+
+- Fix memory leaks in Driver class
+
+- Updated dependencies []:
+  - @ydbjs/core@6.0.6
+
 ## 6.0.5
 
 ### Patch Changes